Refactor more shared application components
