+++ src/areaSearch/actions.js
@@ -21,6 +21,10 @@ import type {
   ReceiveFormValidFlagsAction,
   ReceiveIsSaveClickedAction,
   ShowEditModeAction,
+  FetchListAttributesAction,
+  ReceiveListAttributesAction,
+  ReceiveListMethodsAction,
+  ListAttributesNotFoundAction,
 } from '$src/areaSearch/types';
 import type {Attributes, Methods} from '$src/types';
 import type {
@@ -43,6 +47,18 @@ export const receiveFormValidFlags = (valid: Object): ReceiveFormValidFlagsActio
 
 export const clearFormValidFlags = (): ClearFormValidFlagsAction =>
   createAction('mvj/areaSearch/CLEAR_FORM_VALID_FLAGS')();
+
+export const fetchListAttributes = (): FetchListAttributesAction =>
+  createAction('mvj/areaSearch/FETCH_LIST_ATTRIBUTES')();
+
+export const receiveListAttributes = (attributes: Attributes): ReceiveListAttributesAction =>
+  createAction('mvj/areaSearch/RECEIVE_LIST_ATTRIBUTES')(attributes);
+
+export const receiveListMethods = (methods: Methods): ReceiveListMethodsAction =>
+  createAction('mvj/areaSearch/RECEIVE_LIST_METHODS')(methods);
+
+export const listAttributesNotFound = (): ListAttributesNotFoundAction =>
+  createAction('mvj/areaSearch/LIST_ATTRIBUTES_NOT_FOUND')();
 
 export const fetchAttributes = (): FetchAttributesAction =>
   createAction('mvj/areaSearch/FETCH_ATTRIBUTES')();
