+++ src/main.scss
@@ -76,6 +76,7 @@
 @import "plotApplications/components/exportModal/PlotSearchExportModal";
 @import "areaSearch/components/area-search-application";
 @import "areaSearch/components/map/area-search-selected-area-mini-map";
+@import "src/application/components/applicant-info-check";
 
 @import "foundation/styles";
