+++ src/plotApplications/actions.js
@@ -3,71 +3,67 @@ import {createAction} from 'redux-actions';
 import type {Attributes, Methods} from '$src/types';
 
 import type {
-  FetchPlotApplicationsListAction,
-  ReceivePlotApplicationsListAction,
-  PlotApplicationsList,
-  ReceiveMethodsAction,
-  AttributesNotFoundAction,
-  ReceiveAttributesAction,
-  FetchAttributesAction,
-  PlotApplication,
-  FetchSinglePlotApplicationAction,
-  ReceiveSinglePlotApplicationAction,
-  HideEditModeAction,
-  ShowEditModeAction,
-  ReceiveIsSaveClickedAction,
-  ReceiveCollapseStatesAction,
-  ReceiveFormValidFlagsAction,
-  ClearFormValidFlagsAction,
-  EditPlotApplicationAction,
-  ApplicationsNotFoundAction,
-  FetchPlotApplicationsByBBoxAction,
-  NotFoundByBBoxAction,
-  ReceivePlotApplicationsByBBoxAction,
-  FetchPlotSearchSubtypesAction,
-  PlotSearchSubtypesNotFoundAction,
-  ReceivePlotSearchSubtypesAction,
-  ApplicationRelatedFormNotFoundAction,
-  FetchApplicationRelatedFormAction,
-  ReceiveApplicationRelatedFormAction,
-  FetchApplicationRelatedAttachmentsAction,
-  ReceiveApplicationRelatedAttachmentsAction,
   ApplicationRelatedAttachmentsNotFoundAction,
+  ApplicationRelatedFormNotFoundAction,
+  ApplicationRelatedPlotSearchNotFoundAction,
+  ApplicationsNotFoundAction,
+  AttachmentAttributesNotFoundAction,
+  AttributesNotFoundAction,
+  BatchEditPlotApplicationInfoChecksAction,
+  ClearFormValidFlagsAction,
   CreatePlotApplicationAction,
-  ReceivePlotApplicationSavedAction,
-  ReceivePlotApplicationSaveFailedAction,
-  UploadFileAction,
+  DeleteTargetInfoCheckMeetingMemoAction,
+  DeleteUploadAction,
+  EditPlotApplicationAction,
+  FetchApplicationRelatedAttachmentsAction,
+  FetchApplicationRelatedFormAction,
+  FetchApplicationRelatedPlotSearchAction,
   FetchAttachmentAttributesAction,
+  FetchAttributesAction,
+  FetchPendingUploadsAction,
+  FetchPlotApplicationsByBBoxAction,
+  FetchPlotApplicationsListAction,
+  FetchPlotSearchSubtypesAction,
+  FetchSinglePlotApplicationAction,
+  FetchTargetInfoChecksForPlotSearchAction,
+  HideEditModeAction,
+  InfoCheckBatchEditData,
+  InfoCheckBatchEditErrors,
+  InitializeFormEntriesForApplicationAction,
+  NotFoundByBBoxAction,
+  PendingUploadsNotFoundAction,
+  PlotApplication,
+  PlotApplicationsList,
+  PlotSearchSubtypesNotFoundAction,
+  ReceiveApplicationRelatedAttachmentsAction,
+  ReceiveApplicationRelatedFormAction,
+  ReceiveApplicationRelatedPlotSearchAction,
   ReceiveAttachmentAttributesAction,
   ReceiveAttachmentMethodsAction,
-  AttachmentAttributesNotFoundAction,
-  SetCurrentEditorTargetsAction,
-  FetchApplicationRelatedPlotSearchAction,
-  ReceiveApplicationRelatedPlotSearchAction,
-  ApplicationRelatedPlotSearchNotFoundAction,
+  ReceiveAttributesAction,
+  ReceiveCollapseStatesAction,
   ReceiveFileOperationFinishedAction,
-  DeleteUploadAction,
-  PendingUploadsNotFoundAction,
+  ReceiveFormValidFlagsAction,
+  ReceiveIsSaveClickedAction,
+  ReceiveMethodsAction,
   ReceivePendingUploadsAction,
-  InitializeFormEntriesForApplicationAction,
-  FetchPendingUploadsAction,
-  FetchApplicantInfoCheckAttributesAction,
-  ReceiveApplicantInfoCheckAttributesAction,
-  ApplicantInfoCheckAttributesNotFoundAction,
-  ReceiveUpdatedApplicantInfoCheckItemAction,
-  ReceiveUpdatedTargetInfoCheckItemAction,
-  DeleteTargetInfoCheckMeetingMemoAction,
-  UploadTargetInfoCheckMeetingMemoAction,
-  ReceiveTargetInfoCheckMeetingMemoUploadedAction,
-  TargetInfoCheckMeetingMemoUploadFailedAction,
-  InfoCheckBatchEditData,
-  BatchEditPlotApplicationInfoChecksAction,
-  InfoCheckBatchEditErrors,
-  ReceivePlotApplicationInfoCheckBatchEditSuccessAction,
   ReceivePlotApplicationInfoCheckBatchEditFailureAction,
-  FetchTargetInfoChecksForPlotSearchAction,
+  ReceivePlotApplicationInfoCheckBatchEditSuccessAction,
+  ReceivePlotApplicationSavedAction,
+  ReceivePlotApplicationSaveFailedAction,
+  ReceivePlotApplicationsByBBoxAction,
+  ReceivePlotApplicationsListAction,
+  ReceivePlotSearchSubtypesAction,
+  ReceiveSinglePlotApplicationAction,
+  ReceiveTargetInfoCheckMeetingMemoUploadedAction,
   ReceiveTargetInfoChecksForPlotSearchAction,
+  ReceiveUpdatedTargetInfoCheckItemAction,
+  SetCurrentEditorTargetsAction,
+  ShowEditModeAction,
+  TargetInfoCheckMeetingMemoUploadFailedAction,
   TargetInfoChecksForPlotSearchNotFoundAction,
+  UploadFileAction,
+  UploadTargetInfoCheckMeetingMemoAction,
 } from '$src/plotApplications/types';
 
 export const fetchPlotApplicationsList = (search: string): FetchPlotApplicationsListAction =>
@@ -207,18 +203,6 @@ export const attachmentAttributesNotFound = (): AttachmentAttributesNotFoundActi
 
 export const setCurrentEditorTargets = (payload: Array<Object>): SetCurrentEditorTargetsAction =>
   createAction('mvj/plotApplications/SET_CURRENT_EDITOR_TARGETS')(payload);
-
-export const fetchApplicantInfoCheckAttributes = (): FetchApplicantInfoCheckAttributesAction =>
-  createAction('mvj/plotApplications/FETCH_APPLICANT_INFO_CHECK_ATTRIBUTES')();
-
-export const receiveApplicantInfoCheckAttributes = (payload: Object): ReceiveApplicantInfoCheckAttributesAction =>
-  createAction('mvj/plotApplications/RECEIVE_APPLICANT_INFO_CHECK_ATTRIBUTES')(payload);
-
-export const applicantInfoCheckAttributesNotFound = (): ApplicantInfoCheckAttributesNotFoundAction =>
-  createAction('mvj/plotApplications/APPLICANT_INFO_CHECK_ATTRIBUTES_NOT_FOUND')();
-
-export const receiveUpdatedApplicantInfoCheckItem = (payload: Object): ReceiveUpdatedApplicantInfoCheckItemAction =>
-  createAction('mvj/plotApplications/RECEIVE_UPDATED_APPLICANT_INFO_CHECK_ITEM')(payload);
 
 export const receiveUpdatedTargetInfoCheckItem = (payload: Object): ReceiveUpdatedTargetInfoCheckItemAction =>
   createAction('mvj/plotApplications/RECEIVE_UPDATED_TARGET_INFO_CHECK_ITEM')(payload);
