+++ src/plotApplications/components/infoCheck/PlotApplicationApplicantInfoCheck.js
@@ -6,11 +6,13 @@ import {connect} from 'react-redux';
 import {getFieldOptions, getLabelOfOption} from '$util/helpers';
 import {getUserFullName} from '$src/users/helpers';
 import type {Attributes} from '$src/types';
-import {getApplicantInfoCheckItems} from '$src/plotApplications/helpers';
 import PlotApplicationInfoCheckCollapse from '$src/plotApplications/components/infoCheck/PlotApplicationInfoCheckCollapse';
 import {
-  getApplicantInfoCheckAttributes, getApplicationApplicantInfoCheckData,
+  getApplicationApplicantInfoCheckData,
 } from '$src/plotApplications/selectors';
+import ApplicantInfoCheck from '$src/application/components/infoCheck/ApplicantInfoCheck';
+import {getApplicantInfoCheckAttributes} from '$src/application/selectors';
+import {getApplicantInfoCheckItems} from '$src/application/helpers';
 
 type OwnProps = {
   section: Object,
@@ -32,28 +34,15 @@ class PlotApplicationApplicantInfoCheck extends PureComponent<Props> {
       infoCheckData,
     } = this.props;
 
-    const infoCheckStatusOptions = getFieldOptions(infoCheckAttributes, 'state');
+    const infoCheckStateOptions = getFieldOptions(infoCheckAttributes, 'state');
 
     return (
       <PlotApplicationInfoCheckCollapse className="PlotApplicationApplicantInfoCheck" headerTitle="Hakijan käsittelytiedot">
         <h4>Tarkistettavat dokumentit</h4>
-        <Row>
-          {infoCheckData.map((item) => {
-            const statusText = getLabelOfOption(infoCheckStatusOptions, item.data.state);
-
-            return <Column small={6} key={item.kind.type}>
-              <Row>
-                <Column small={8}>
-                  <span>{item.kind.label}</span>
-                </Column>
-                <Column small={4}>
-                  {statusText}
-                  {item.data.preparer && <>, {getUserFullName(item.data.preparer)}</>}
-                </Column>
-              </Row>
-            </Column>;
-          })}
-        </Row>
+        <ApplicantInfoCheck
+          infoChecks={infoCheckData}
+          infoCheckStateOptions={infoCheckStateOptions}
+        />
       </PlotApplicationInfoCheckCollapse>
     );
   }
