+++ src/plotApplications/components/infoCheck/_plotApplicationInfoCheckCollapse.scss
@@ -1,18 +1,4 @@
 .PlotApplicationInfoCheckCollapse {
-  .collapse__content {
-    &, span, a {
-      font-size: $font-small;
-    }
-  }
-
-  &.PlotApplicationApplicantInfoCheck {
-    .collapse__content {
-      .row {
-        line-height: 1.5rem;
-      }
-    }
-  }
-
   .modal__wrapper {
     overflow: visible;
   }
@@ -22,11 +8,4 @@
   // Fix cut-off dropdowns
   // The original rule in the collapse component is itself marked as important.
   overflow: visible !important;
-}
-
-
-.PlotApplicationApplicantInfoCheckEdit__item--dirty {
-  .column:nth-child(2) > * {
-    background-color: darken($is-dirty-color, 5%);
-  }
 }
