+++ src/root/types.js
@@ -42,10 +42,12 @@ import type {VatState} from '$src/vat/types';
 import type {LandUseInvoicesState} from '$src/landUseInvoices/types';
 import type {LandUseAgreementAttachmentState} from '$src/landUseAgreementAttachment/types';
 import type {AreaSearchState} from '$src/areaSearch/types';
+import type {ApplicationState} from '$src/application/types';
 
 export type RootState = {
   api: ApiState,
   apiToken: AuthState,
+  application: ApplicationState,
   areaNote: AreaNoteState,
   areaSearch: AreaSearchState,
   auditLog: AuditLogState,
